Cleaning Windows compilation by adding headers hack for debug config. Additional: * reduced tests duration by limiting tests to 100 iterations * removed legacy tests

+++ tests/test_utils.cpp
@@ -39,7 +39,7 @@ BOOST_AUTO_TEST_CASE(test_report)
     double max_L2 = std::numeric_limits<double>::lowest();
     for (size_t i = 0; i < (rng() % 100); i++)
     {
-        double rand_L2 = rng() % 10000 - 5000;
+        double rand_L2 = static_cast<double>(rng() % 10000 - 5000);
         results.vec_L2.emplace_back(rand_L2);
         max_L2 = std::max(max_L2, rand_L2);
     }
